Major refactor, cleanup and tests
