+++ .travis.yml
@@ -5,6 +5,6 @@ go:
     - 1.8
     - tip
 before_install:
-    - go get -u github.com/golang/lint/golint
+    - make gettools
 script:
     - make
